+++ api-builder-plugin-fn-file/src/actions.js
@@ -1,6 +1,0 @@
-const { readCVSFile } = require('./csvFileActions')
-const { writeFile } = require('./standardFileActions')
-
-module.exports = {
-	readCVSFile, writeFile
-};